Issue 2 & 7 from List TODO Solved and BugFix
Issue 2 - The game nos starts with the control panel opened
Issue 7 - The planet orbits now disappear when the planet is destroyed
BugFix - There was an issue with the rotation of the camera target. When
a planet was destroyed and we change the camera target, the camera was
still targeting destroyed planets.